make single header; fix g++11 errors
